% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{Global coronavirus statistics}
\format{
A data frame that consists of 65,330 observations and 8 variables.
\itemize{
\item \strong{Date_reported}: dates from 2020-01-03 to 2020-10-06
\item \strong{Country_code}: two letter country code
\item \strong{Country}: name of country
\item \strong{WHO_region}: 4 and 5 letter codes that WHO has designated to divide the world into regions
\item \strong{New_cases}: number of daily cases reported on a specific date
\item \strong{Cumulative_cases}: total cases reported
\item \strong{New_deaths}: number of daily deaths reported on a specific date
\item \strong{Cumulative_deaths}: total deaths reported
}
}
\source{
\href{https://covid19.who.int/}{World Health Organization}
}
\usage{
data2
}
\description{
A dataset displaying the changes in cumulative cases and deaths over time for 235 countries.
}
\keyword{datasets}
