% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{data3}
\alias{data3}
\title{Global summary coronavirus statistics}
\format{
A data frame that consists of 235 observations and 3 variables.
\itemize{
\item \strong{Country}: name of country
\item \strong{Cumulative_cases}: total cases reported
\item \strong{Cumulative_deaths}: total deaths reported
}
}
\source{
\href{https://covid19.who.int/}{World Health Organization}
}
\usage{
data3
}
\description{
A dataset showing the cumulative cases and deaths for 235 countries as at 6th October 2020.
}
\keyword{datasets}
