% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data5.R
\docType{data}
\name{data5}
\alias{data5}
\title{Australia coronavirus cumulative case statistics}
\format{
A data frame that consists of 2080 observations and 5 variables.
\itemize{
\item \strong{State}: name of the Australian state
\item \strong{Lat}: the latitude coordinate for the Australian state
\item \strong{Long}: the longitude coordinate for the Australian state
\item \strong{Date}: date on which cases were reported
\item \strong{Cumulative Cases}: the total number of cases reported
}
}
\source{
\href{https://data.humdata.org/dataset/novel-coronavirus-2019-ncov-cases#}{HDX}
}
\usage{
data5
}
\description{
A dataset displaying the changes in cumulative cases in all Australian states over a time period of January to October.
}
\keyword{datasets}
