% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launch_app}
\alias{launch_app}
\title{Launch a shiny app}
\usage{
launch_app()
}
\value{
A shiny app about coronavirus statistics around the world
}
\description{
This function launches a shiny app when executed by the user
}
\examples{
\dontrun{
launch_app()
}

}
