% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafserv.R
\name{leafserv}
\alias{leafserv}
\title{Generate a leaflet map for a country}
\usage{
leafserv(Long, Lat, State)
}
\arguments{
\item{Long}{The longitude coordinate of the country.}

\item{Lat}{The latitude coordinate of the country.}

\item{State}{The state(s) within the country. This appears as a popup on the map.}
}
\value{
A leaflet map with the outline of the desired country is displayed. On the map, the longitude and latitude coordinates of the states are plotted.
}
\description{
This function is used to create a leaflet map using longitude, latitude coordinates and state names from a dataframe. By parsing this information into the leafserv() function, a leaflet map of the desired country will be generated.
}
\examples{
\dontrun{
leafserv(df$Long, df$Lat, df$State)
}

}
