% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectui.R
\name{selectui01}
\alias{selectui01}
\alias{selectui02}
\alias{selectui03}
\title{Create functions to simplify repetitive calls for selectInputs in the user interface-side of a shiny app}
\usage{
selectui01(inputId)
}
\arguments{
\item{inputId}{An id that is unique in the user interface-side in order for the server side of the shiny app to call on.}
}
\value{
A selectInput option
}
\description{
Create functions to simplify repetitive calls for selectInputs in the user interface-side of a shiny app
}
\examples{
inputId <- state
selectui01("state")
inputId <- countriesp
selectui02("countriesp")
inputId <- country
selectui03("country")

}
