% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectui.R
\name{selectui01}
\alias{selectui01}
\alias{selectui02}
\alias{selectui03}
\title{Simplify repetitive calls for selectInputs in the user interface-side of a shiny app}
\usage{
selectui01(inputId)
}
\arguments{
\item{inputId}{An id that is unique in the user interface-side in order for the server side of the shiny app to call on.}
}
\value{
A selectInput option
}
\description{
This function is used to display a drop-down list of options for a user to select either states/countries to visualize coronavirus statistics.
}
\examples{
inputId <- state
selectui01("state")
inputId <- country
selectui03("country")

}
